Fixed unicode problem.
Problem with colons in file names was that the file system didn't like
it & gave it a weird encoding.  Was able to convert back in Linux, but
Windows didn't like the unicode in the python code.  Fixed it by
converting the colons to underscores in the file names & converting
back to colons for dates.