6/30 data update
This update includes an increase in the number of NYC residents who died due to COVID-19. Most of this increase is due to new information we received from the NYS Department of Health about NYC residents who died outside the city. The vast majority of these deaths occurred over three weeks ago.